Split test cases into seperate files
